warnings
Former-commit-id: eb23d8a8fe1e54ec72104335494287521bf8d79f

+++ modules/DataLayer/Storage/BinarySerialization.h
@@ -17,7 +17,7 @@ T bswap( T w )
   
   // decent compilers will unroll this (gcc)
   // or even convert straight into single bswap (clang)
-  for (int i = 0; i < sizeof(r); i++) 
+  for (size_t i = 0; i < sizeof(r); i++) 
     {
     r <<= 8;
     r |= w & 0xff;
@@ -43,7 +43,7 @@ T endian(T w, Endianness endian )
   
   // decent compilers will unroll this (gcc)
   // or even convert straight into single bswap (clang)
-  for (int i = 0; i < sizeof(r); i++) {
+  for (size_t i = 0; i < sizeof(r); i++) {
   r <<= 8;
   r |= w & 0xff;
   w >>= 8;
